+++ ApiGympass/Data/Repositories/Interfaces/IGymRepository.cs
@@ -7,7 +7,7 @@ namespace ApiGympass.Data.Repositories.Interfaces
         Task<Gym> CreateGymAsync(Gym gym);
         Task<Gym?> FindById(Guid gymId);
         Task<IEnumerable<Gym>> SearchManyAsync(string query, int page, int pageSize);
-        Task<IEnumerable<Gym>> FindManyNearbyAsync(double latitude, double longitude);
+        Task<IEnumerable<Gym>> FindManyNearbyAsync(decimal latitude, decimal longitude);
         Task<int> CountAsync(string query);
 
     }
